% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CutSeq.R
\name{CutSeq}
\alias{CutSeq}
\alias{CutSeqEqu}
\title{Cut vectors}
\usage{
CutSeq(cutSeq)

CutSeqEqu(vecLen, equNum)
}
\arguments{
\item{cutSeq}{The inverals vector. The length of cutSeq could be more than 1, and 0 will be automatically excluded.}

\item{vecLen}{The length of vector used to cut.}

\item{equNum}{The equal internal.}
}
\value{
A cut matrix in which the first row is the start point and second row is the end point.
}
\description{
Cut vectors with invervals
}
\details{
CutSeq() is used to cut a vector with different invervals. CutSeqEqu() is used to cut a vector with same invervals.
}
\examples{
# with one interval
CutSeq(10)
# with multiple interval
CutSeq(c(2, 3, 5))
# exclude 0
# equal interval is the same as the length of vector
CutSeqEqu(10, equNum = 10)
CutSeqEqu(21, equNum = 10)
# euqal interval is larger than the length of vector
CutSeqEqu(10, equNum = 20)
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}

